Run everything with sudo.
Since the pack script needs to be run with sudo for ensuring files download/execution/edit, it won't be able to use the user-installed dotnet sdk (i.e. 5.0).
The solution is to not use the setup-dotnet github action, and install it manually, taking care of running everything else with sudo.
This workaround can be removed when github actions will default to net5.0 instead of netcoreapp3.1.